requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Babel==3.1.0
Flask-WTF==1.2.1
Flask-Caching>=2.0.0
Flask-Limiter>=3.0.0
email-validator==2.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pyfluidsynth==1.3.2
mido==1.3.0
numpy==1.26.0
scipy==1.11.4
librosa==0.10.1
Babel==2.13.1
Jinja2==3.1.2
itsdangerous==2.1.2
SQLAlchemy==2.0.23
alembic==1.12.1
click==8.1.7
Flask-HTTPAuth>=4.7.0
Flask-ReCAPTCHA==0.4.2
pretty_midi>=0.2.10
midi2audio>=0.1.1
pydub>=0.25.1